split apart war project and jetty-standalone project
this removes the need for a custom ant script that filters dependencies and classes in the war file build
